feat: rework test
