+++ test/vectors_test.dart
@@ -1,0 +1,38 @@
+import 'dart:core';
+import 'package:http/http.dart' as http;
+import 'package:bip340/bip340.dart' as bip340;
+import 'package:csv/csv.dart' as csv;
+import 'package:test/test.dart';
+
+void main() async {
+  // test bip340 vectors from the bip itself
+  var data = await http.read(Uri.parse(
+      "https://raw.githubusercontent.com/bitcoin/bips/master/bip-0340/test-vectors.csv"));
+  List<List<dynamic>> vectors = csv.CsvToListConverter().convert(
+    data,
+    shouldParseNumbers: false,
+  );
+
+  for (var i = 1; i < vectors.length; i++) {
+    group('Vector $i', () {
+      var line = vectors[i];
+      var errorMessage = line[7].length == 0 ? "<blank>" : line[7];
+
+      if (line[3].length > 0) {
+        test('Sign $errorMessage', () {
+          var sig = bip340.sign(line[1], line[4], line[3]);
+          var expected = line[5].toLowerCase();
+
+          expect(sig, expected);
+        });
+      }
+
+      test('Verification $errorMessage', () {
+        var ok = bip340.verify(line[2], line[4], line[5]);
+        var expected = line[6].toLowerCase();
+
+        expect(ok.toString(), expected);
+      });
+    });
+  }
+}
